Converted index.js -> index.ts and added pathways-model declaration file.
