+++ src/engine/elm_executor.js
@@ -4,7 +4,7 @@ const cqlfhir = require('cql-exec-fhir');
 
 /**
  * Engine function that takes in a patient file (JSON) and an ELM file, running the patient against the ELM file
- * @param {string} patient - JSON string representing the patient's health record
+ * @param patient - JSON string representing the patient's health record
  * @param {string} elm - ELM string representing the ELM (previosuly converted from CQL) on which the patient will be run.
  * @return {object} returns a JSON object which is the result of analyzing the patient against the elm file
  */
