+++ src/engine/output_results.js
@@ -2,7 +2,7 @@
 /* eslint-disable max-len */
 /**
  * Engine function to take in the ELM patient results and output data relating to the patient's pathway
- * @param {string} pathway - JSON (or string representing) the entire pathway
+ * @param pathway - JSON (or string representing) the entire pathway
  * @param {string} patientData - JSON (or string representing) the data on the patient from a CQL execution. Note this is a single patient not the entire patientResults object
  * @return {object} returns a JSON object describing where the patient is on the given pathway
  *  {
